Refactor layout-specific portions of probe
