+++ presto-main/src/main/java/com/facebook/presto/operator/HashGenerator.java
@@ -14,7 +14,6 @@
 package com.facebook.presto.operator;
 
 import com.facebook.presto.spi.Page;
-import com.facebook.presto.spi.block.BlockDecoder;
 
 import static com.google.common.base.Preconditions.checkState;
 
@@ -33,10 +32,5 @@ public interface HashGenerator
 
         checkState(partition >= 0 && partition < partitionCount);
         return partition;
-    }
-
-    default void getPartitions(int partitionCount, Page page, BlockDecoder decoder, int[] partitionsOut)
-    {
-        throw new UnsupportedOperationException();
     }
 }
