+++ presto-main/src/main/java/com/facebook/presto/operator/InterpretedHashGenerator.java
@@ -16,9 +16,6 @@ package com.facebook.presto.operator;
 import com.facebook.presto.operator.scalar.CombineHashFunction;
 import com.facebook.presto.spi.Page;
 import com.facebook.presto.spi.block.Block;
-import com.facebook.presto.spi.block.BlockDecoder;
-import com.facebook.presto.spi.block.LongArrayBlock;
-import com.facebook.presto.spi.type.AbstractLongType;
 import com.facebook.presto.spi.type.Type;
 import com.facebook.presto.sql.planner.optimizations.HashGenerationOptimizer;
 import com.facebook.presto.type.TypeUtils;
@@ -64,52 +61,6 @@ public class InterpretedHashGenerator
             result = CombineHashFunction.getHash(result, TypeUtils.hashPosition(type, blockProvider.apply(hashChannels[i]), position));
         }
         return result;
-    }
-
-    @Override
-    public void getPartitions(int partitionCount, Page page, BlockDecoder decoder, int[] partitionsOut)
-    {
-        int positionCount = page.getPositionCount();
-        long[] hashes = null;
-        synchronized (this) {
-            // Access only once. Another thread may set hashesInReserve wihout synchronization.
-            long[] candidate = hashesInReserve;
-            if (candidate != null && candidate.length >= positionCount) {
-                hashes = candidate;
-                hashesInReserve = null;
-            }
-        }
-        if (hashes == null) {
-            hashes = new long[positionCount];
-        }
-        for (int position = 0; position < positionCount; position++) {
-            hashes[position] = HashGenerationOptimizer.INITIAL_HASH_VALUE;
-        }
-        for (int i = 0; i < hashChannels.length; i++) {
-            Type type = hashChannelTypes.get(i);
-            Block block = page.getBlock(i);
-            decoder.decodeBlock(block);
-            Block leafBlock = decoder.getLeafBlock();
-            if (leafBlock instanceof LongArrayBlock) {
-                long[] longs = decoder.getValues(long[].class);
-                int[] longsMap = decoder.getRowNumberMap();
-                boolean[] nulls = decoder.getValueIsNull();
-                for (int position = 0; position < positionCount; position++) {
-                    int valueIdx = longsMap[position];
-                    hashes[position] = CombineHashFunction.getHash(hashes[position],
-                            (nulls == null || !nulls[valueIdx])
-                                    ? AbstractLongType.hash(longs[valueIdx])
-                                    : TypeUtils.NULL_HASH_CODE);
-                }
-            }
-            else {
-                for (int position = 0; position < positionCount; position++) {
-                    hashes[position] = CombineHashFunction.getHash(hashes[position], TypeUtils.hashPosition(type, block, position));
-                }
-            }
-            decoder.release();
-        }
-        hashesInReserve = hashes;
     }
 
     @Override
