+++ presto-main/src/main/java/com/facebook/presto/operator/PrecomputedHashGenerator.java
@@ -14,8 +14,6 @@
 package com.facebook.presto.operator;
 
 import com.facebook.presto.spi.Page;
-import com.facebook.presto.spi.block.Block;
-import com.facebook.presto.spi.block.BlockDecoder;
 import com.facebook.presto.spi.type.BigintType;
 
 import static com.google.common.base.MoreObjects.toStringHelper;
@@ -34,27 +32,6 @@ public class PrecomputedHashGenerator
     public long hashPosition(int position, Page page)
     {
         return BigintType.BIGINT.getLong(page.getBlock(hashChannel), position);
-    }
-
-    @Override
-    public void getPartitions(int partitionCount, Page page, BlockDecoder decoder, int[] partitionsOut)
-    {
-        Block block = page.getBlock(hashChannel);
-        decoder.decodeBlock(block);
-        int positionCount = block.getPositionCount();
-        long[] longs = decoder.getValues(long[].class);
-        if (decoder.isIdentityMap()) {
-            for (int i = 0; i < positionCount; i++) {
-                partitionsOut[i] = (int) ((longs[i] & 0x7fffffffffffL) % partitionCount);
-            }
-        }
-        else {
-            int[] map = decoder.getRowNumberMap();
-            for (int i = 0; i < positionCount; i++) {
-                partitionsOut[i] = (int) ((longs[map[i]] & 0x7fffffffffffL) % partitionCount);
-            }
-        }
-        decoder.release();
     }
 
     @Override
